backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
redis==5.0.1
celery==5.3.4
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.25.2
pandas==2.1.4
scikit-learn==1.3.2
lightgbm==4.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
boto3==1.34.0
stripe==7.8.0
twilio==8.11.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
black==23.11.0
isort==5.13.0
flake8==6.1.0
# Additional dependencies for proper functioning
typing-extensions==4.8.0
email-validator==2.1.0
